# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a proof-of-concept to describe the idea depicted in issue [livepeer/research/issues/36](https://github.com/livepeer/research/issues/36)",
        "version": "0.0.0",
        "title": "Content Authenticity Verification",
        "contact": {
            "email": "michael@livepeer.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:8088",
    "_basePath": "/v2",
    "tags": [
        {
            "name": "ffmpeg",
            "description": "demonstration of ffmpeg interaction",
            "externalDocs": {
                "description": "Using macroblocks from three separate transcode rendition jobs to determine content authenticity",
                "url": "https://github.com/livepeer/research/issues/36"
            }
        }
    ],
    "schemes": [
        "http"
    ],
    "paths": {
        "/scale": {
            "get": {
                "tags": [
                    "ffmpeg"
                ],
                "summary": "extract still image and scale",
                "description": "extract still image and scale",
                "operationId": "extractScale",
                "produces": [
                    "image/jpeg"
                ],
                "parameters": [
                    {
                        "name": "s",
                        "in": "query",
                        "description": "Source video file to extract a still image from",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "w",
                        "in": "query",
                        "description": "Desired width of the extracted image",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "h",
                        "in": "query",
                        "description": "Desired height of the extracted image",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "o",
                        "in": "query",
                        "description": "Offset (in seconds) from the start of the video from which image shall be extracted",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": "image/*",
                        "schema": {
                            "type": "string",
                            "format": "binary"
                        }
                    },
                    "400": {
                        "description": "Bad request params"
                    },
                    "500": {
                        "description": "Task error"
                    }
                }
            }
        },
        "/bitmap": {
            "get": {
                "tags": [
                    "ffmpeg"
                ],
                "summary": "extract still image bitmap and scale",
                "description": "extract still image bitmap and scale",
                "operationId": "extractScale",
                "produces": [
                    "image/bmp"
                ],
                "parameters": [
                    {
                        "name": "s",
                        "in": "query",
                        "description": "Source video file to extract a still image bitmap from",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "w",
                        "in": "query",
                        "description": "Desired width of the extracted image bitmap",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "h",
                        "in": "query",
                        "description": "Desired height of the extracted image bitmap",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "o",
                        "in": "query",
                        "description": "Offset (in seconds) from the start of the video from which image bitmap shall be extracted",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": "image/*",
                        "schema": {
                            "type": "string",
                            "format": "binary"
                        }
                    },
                    "400": {
                        "description": "Bad request params"
                    },
                    "500": {
                        "description": "Task error"
                    }
                }
            }
        }
    },
    "externalDocs": {
        "description": "Find out more about Swagger",
        "url": "http://swagger.io"
    }
}